+++ CHANGES.rst
@@ -7,6 +7,7 @@ What's new
 * Added support for target masks when regridding ``LocStream`` to ``Grid`` with ``nearest_s2d`` (:pull:`445`). By `Martin Schupfner <https://github.com/sol1105>`_.
 * ``xesmf.util.cf_grid_2d`` returns bounds as coordinates, as ``grid_2d`` does and as usually expected. (:pull:`453`). By `Pascal Bourgault <https://github.com/aulemahal>`_.
 * Accept masks in "X, Y" order (:issue:`447`, :pull:`456`). By `Pascal Bourgault <https://github.com/aulemahal>`_.
+* Correctly use arguments ``input_dims`` and ``output_dims`` in ``Regridder`` when grids are defined as dictionaries of numpy arrays (:issue:`362`, :pull:`455`). By `Aaron G Meyer <https://github.com/agmeyer4>`_.
 
 0.8.10 (2025-04-29)
 -------------------